fix: resolve DYNAMIC_SERVER_USAGE and invalid deploy workflow
- deploy.yml: add required 'jobs' key (was failing GitHub Actions validation)
- catalog/page.tsx: convert to Server Component with Suspense boundary
- catalog/CatalogContent.tsx: extract client component (uses useSearchParams)
- games/guess-the-flag/page.tsx: convert to Server Component with Suspense boundary
- games/guess-the-flag/GameContent.tsx: extract client component (uses useSearchParams)

Fixes FUNCTION_INVOCATION_FAILED on Vercel caused by useSearchParams()
being called inside a Client Component Suspense boundary instead of a
Server Component one.
