+++ src/app/[locale]/games/guess-the-flag/GameContent.tsx
@@ -1,0 +1,28 @@
+"use client";
+
+import { useMemo } from "react";
+import { useCountries } from "@/lib/providers/CountriesProvider";
+import { filterByContinent } from "@/lib/utils/countries";
+import { useContinentFilter } from "@/lib/hooks/useContinentFilter";
+import { useUserProgress } from "@/lib/hooks/useUserProgress";
+import { AppShell } from "@/components/layout/AppShell";
+import { FlagQuiz } from "@/components/games/FlagQuiz";
+
+export function GameContent() {
+  const { countries } = useCountries();
+  const { activeContinent } = useContinentFilter();
+  const { saveQuizScore } = useUserProgress();
+
+  const pool = useMemo(
+    () => filterByContinent(countries, activeContinent),
+    [countries, activeContinent],
+  );
+
+  return (
+    <AppShell>
+      <div className="max-w-4xl mx-auto">
+        <FlagQuiz pool={pool} onGameOver={saveQuizScore} />
+      </div>
+    </AppShell>
+  );
+}
